refactor: moved lib folder to shared
