+++ src/actions/category/categoryOptions.ts
@@ -3,13 +3,8 @@
 import { OptionSetType } from "@prisma/client";
 import { z } from "zod";
 
-import { db } from "@/lib/db";
-import {
-  TOptionSet,
-  TSingleOption,
-  TSingleSpec,
-  TSpecGroup,
-} from "@/types/common";
+import { db } from "@/shared/lib/db";
+import { TOptionSet, TSingleOption, TSingleSpec, TSpecGroup } from "@/types/common";
 
 const AddOptionSet = z.object({
   name: z.string().min(3),
