+++ src/app/admin/layout.tsx
@@ -3,8 +3,7 @@ import { redirect } from "next/navigation";
 import { getServerSession } from "next-auth";
 
 import AdminSidebar from "@/components/admin/sideBar";
-import { authOptions } from "@/lib/authOptions";
-
+import { authOptions } from "@/shared/lib/authOptions";
 
 export const metadata: Metadata = {
   title: "Admin",
