+++ src/shared/lib/authOptions.ts
